chore: update package dependencies and remove unused TaskDashboard and RadicalChart components
